move rewards module to trait
